requirements.txt
python-dotenv
python-binance
websocket_client
numpy
simple_chalk
ta
pandas
pandas-ta